feat (page): added MyAnimeList Integration
This integration allows users to add and update anime on their MAL,
straight from 9anime. This integration can be enabled by going to the
settings page. Once enabled, MAL username and password will have to
be entered by going to the options.

Screen-shot:
-ADD: https://image.ibb.co/eBj75k/mal_add.png
-UPDATE: https://image.ibb.co/jojUJ5/mal_update.png

Disclaimer:
The username and password are stored in plain text in the browsers local
storage.