feat (download_all): add support for external downloaders
This is the initial implementation of external downloaders functionality
for Download All. It fetches file links at 2 seconds interval and once all the
links are fetched a popup is shown to the user.
(https://image.ibb.co/iRqP5k/external_downloads.png)

Please Note:
There is an issue with IDM where in it does not resolve some of the file
URL's (https://image.ibb.co/mC3cQk/idm_error.png). The only
explanation I can think of is that google does not like IDM resolving that
many files in such a short span. This needs to be thoroughly tested.

This marks the initial implementation of #8.